JBRULES-1498 Thead safe partitioning of WorkingMemory entry points -currently id's are int's, not longs. So need to cast.
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@18817 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-jsr94/src/main/java/org/drools/jsr94/rules/Jsr94FactHandle.java
@@ -77,7 +77,7 @@ public class Jsr94FactHandle extends DefaultFactHandle
     Jsr94FactHandle(final long id,
                     final Object object,
                     final long recency) {
-        super( id,
+        super( (int) id,
                object,
                recency );
     }
